Fix Tiny Tina's Wonderlands code parsing for all platforms
The Tiny Tina's Wonderlands page (http://orcz.com/Tiny_Tina%27s_Wonderlands:_Shift_Codes) has a different format for its SHiFT codes than the other Borderlands pages (e.g. http://orcz.com/Borderlands_2:_Golden_Key) due to its codes being universal. Instead of listing a code for each platform, it lists a code and then whether it works on each platform.

For the moment, these codes work on all platforms, so if a code is present and non-expired in a row, we can return that the code is compatible for all platforms.